+++ Gateway/routes/class-service-routes.php
@@ -6,7 +6,7 @@ use Illuminate\Support\Facades\Route;
 Route::middleware(['auth:api', 'force.pwd'])->prefix(config('gateway.class_prefix'))->group(function(){
     Route::get('all', [ClassController::class, 'index']);
     Route::get('/', [ClassController::class, 'show']);
-    Route::post('/', [ClassController::class, 'store'])->middleware('check.role:SuperAdmin,Admin');
-    Route::post('update', [ClassController::class, 'update'])->middleware('check.role:SuperAdmin,Admin');
-    Route::delete('/{id}', [ClassController::class, 'destroy'])->middleware('check.role:SuperAdmin,Admin');
+    Route::post('/', [ClassController::class, 'store'])->middleware('check.role:SuperAdmin,Admin,AdminSekolah');
+    Route::post('update', [ClassController::class, 'update'])->middleware('check.role:SuperAdmin,Admin,AdminSekolah');
+    Route::delete('/{id}', [ClassController::class, 'destroy'])->middleware('check.role:SuperAdmin,Admin,AdminSekolah');
 });
